Move commands to be stored at root for all cases
