+++ cmd/main.go
@@ -16,6 +16,7 @@ package main
 import (
 	"fmt"
 	"github.com/pufferpanel/pufferpanel/v2"
+	"github.com/spf13/cobra"
 	"os"
 	"runtime/debug"
 )
@@ -36,4 +37,26 @@ func main() {
 	}()
 
 	Execute()
+}
+
+var rootCmd = &cobra.Command{
+	Use:   "pufferpanel",
+	Short: "Game Server Management Panel",
+}
+
+func init() {
+	rootCmd.AddCommand(
+		runCmd,
+		versionCmd,
+		userCmd,
+		shutdownCmd,
+		migrateCmd)
+}
+
+func Execute() {
+	rootCmd.SetVersionTemplate(pufferpanel.Display)
+
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println(err)
+	}
 }
